Implements ExperimentConfiguration creation
ExperimentConfiguration objects are now created from DiffractionPlans.
This is not implemented correctly, since DiffractionPlans is not
implemented as it was specified (see MXSW-1115). There are several
FIXMEs to indicate where changes need to be made.

I have added a FIXME to the ScanMetadata creation method to try to make
clear what it needs to do.

Signed-off-by: Michael Wharmby <michael.wharmby@diamond.ac.uk>